Rename `recycle.js` to `compare.js`
